app/langchain/initialize/embedModel.py
import getpass
import os
from dotenv import load_dotenv
load_dotenv()
if not os.getenv("NVIDIA_API_KEY"):
  os.environ["NVIDIA_API_KEY"] = getpass.getpass("Enter API key for NVIDIA: ")

from langchain_nvidia_ai_endpoints import NVIDIAEmbeddings

embeddings = NVIDIAEmbeddings(model="NV-Embed-QA")