requirements.txt
fastapi
uv
fastapi[standard]
uvicorn[standard]
python-dotenv 
supabase
langchain-text-splitters
langchain-community
langgraph
langchain[together]
langchain-nvidia-ai-endpoints
langchain-astradb
numpy==1.26.4
astrapy
pypdf
pillow
rapidocr-onnxruntime
python-multipart